add a settings page
